+++ app/src/pages/Login.js
@@ -16,9 +16,10 @@ import {
   setupWeb3,
   hasMetamask,
   signMessage
-} from './web3Helpers'
+} from '../web3Helpers'
 import AuthForm from './AuthForm'
-import './login.css'
+import metamaskDownloadImg from './../assets/download-metamask.png'
+import './Login.css'
 
 const propTypes = {
   user: PropTypes.shape({
@@ -39,21 +40,28 @@ export const Login = ({
   authWithSAN,
   client
 }) => {
-  if (user.data.username && user.data.username.length > 0) {
-    return <Redirect to={'/'} />
+  if (user.data.hasOwnProperty('username')) {
+    return <Redirect to='/' />
   }
   return (
     <div className='page wrapper'>
-      <div className='loginContainer'>
+      <div className='login-container'>
         {user.isLoading && !user.hasMetamask && <div>Loading</div>}
         {!user.hasMetamask && !user.isLoading &&
-          <Message
-            warning
-            header={'We can\'t detect Metamask!'}
-            list={[
-              'We can auth you with Metamask account. It\'s secure and easy.'
-            ]}
-          />}
+          <Message warning>
+            <h4>We can't detect Metamask!</h4>
+            <p>We can auth you with Metamask account. It's secure and easy.</p>
+            <div className='help-links'>
+              <a
+                target='_blank'
+                rel='noopener noreferrer'
+                href='https://metamask.io/#how-it-works'>How Metamask works?</a>
+              <a href='https://metamask.io/'>
+                <img width={128} src={metamaskDownloadImg} alt='Metamask link' />
+              </a>
+            </div>
+          </Message>
+        }
         {user.hasMetamask && !user.token &&
           <AuthForm
             account={user.account}
@@ -101,7 +109,7 @@ const mapDispatchToProps = dispatch => {
             type: 'FAILED_LOGIN',
             errorMessage: error
           })
-          console.error(error)
+          throw new Error(error)
         })
       }).catch(error => {
         // TODO: 2017-12-05 16:05 | Yura Zatsepin:
@@ -129,11 +137,11 @@ const mapDispatchToProps = dispatch => {
 }
 
 const requestAuthGQL = gql`
-  mutation ethLogin($signature: String, $address: String, $addressHash: String) {
+  mutation ethLogin($signature: String!, $address: String!, $messageHash: String!) {
     ethLogin(
       signature: $signature,
       address: $address,
-      addressHash: $addressHash) {
+      messageHash: $messageHash) {
         token,
         user {
           id,