+++ config/config.exs
@@ -31,6 +31,11 @@ config :logger, :console,
   handle_otp_reports: true,
   handle_sasl_reports: true
 
+# Error tracking
+config :sentry,
+  included_environments: [:prod],
+  environment_name: Mix.env
+
 config :sanbase, Sanbase.Prices.Store,
   host: {:system, "INFLUXDB_HOST", "localhost"},
   port: {:system, "INFLUXDB_PORT", 8086},
@@ -42,6 +47,15 @@ config :sanbase, Sanbase.Github.Store,
   port: {:system, "INFLUXDB_PORT", 8086},
   pool: [ max_overflow: 10, size: 20 ],
   database: "github_activity"
+
+config :sanbase, SanbaseWorkers.ImportGithubActivity,
+  s3_bucket: {:system, "GITHUB_ARCHIVE_BUCKET", "santiment-github-archive"}
+
+config :sanbase, Sanbase.ExternalServices.TwitterData.Store,
+  host: {:system, "INFLUXDB_HOST", "localhost"},
+  port: {:system, "INFLUXDB_PORT", 8086},
+  pool: [ max_overflow: 10, size: 20 ],
+  database: "twitter_followers_data"
 
 config :hammer,
   backend: {Hammer.Backend.ETS, [expiry_ms: 60_000 * 60 * 4,
@@ -86,8 +100,23 @@ config :sanbase, Sanbase.ExternalServices.Etherscan.Worker,
   update_interval: 5 * 1000 * 60, # 5 minutes
   sync_enabled: {:system, "ETHERSCAN_CRAWLER_ENABLED", false}
 
+config :sanbase, Sanbase.ExternalServices.Github,
+  update_interval: 60 * 1000 * 60, # 60 minutes
+  sync_enabled: {:system, "GITHUB_SCHEDULER_ENABLED", false}
+
 config :sanbase, Sanbase.ExternalServices.Etherscan.Requests,
   apikey: {:system, "ETHERSCAN_APIKEY"}
+
+config :sanbase, Sanbase.ExternalServices.TwitterData.Worker,
+  consumer_key: {:system, "TWITTER_CONSUMER_KEY"},
+  consumer_secret: {:system, "TWITTER_CONSUMER_SECRET"},
+  update_interval: 1000 * 60 * 60 * 6, # 6 hours
+  sync_enabled: {:system, "TWITTER_SCRAPER_ENABLED", false}
+
+config :sanbase, Sanbase.ExternalServices.TwitterData.HistoricalData,
+  apikey: {:system, "TWITTERCOUNTER_API_KEY"},
+  update_interval: 1000 * 60 * 60 * 24, # 1 day
+  sync_enabled: {:system, "TWITTERCOUNTER_SCRAPER_ENABLED", false}
 
 config :sanbase, Sanbase.Notifications.CheckPrice,
   webhook_url: {:system, "NOTIFICATIONS_WEBHOOK_URL"},
@@ -97,10 +126,15 @@ config :sanbase, SanbaseWeb.Guardian,
   issuer: "santiment",
   secret_key: {SanbaseWeb.Guardian, :get_config, [:secret_key_base]}
 
-config :sanbase, Sanbase.Auth.Ethauth,
+config :sanbase, Sanbase.InternalServices.Ethauth,
   url: {:system, "ETHAUTH_URL"},
   basic_auth_username: {:system, "ETHAUTH_BASIC_AUTH_USERNAME"},
   basic_auth_password: {:system, "ETHAUTH_BASIC_AUTH_PASSWORD"}
+
+config :sanbase, Sanbase.InternalServices.Parity,
+  url: {:system, "PARITY_URL"},
+  basic_auth_username: {:system, "PARITY_BASIC_AUTH_USERNAME"},
+  basic_auth_password: {:system, "PARITY_BASIC_AUTH_PASSWORD"}
 
 config :faktory_worker_ex,
   host: {:system, "FAKTORY_HOST", "localhost"},
@@ -113,6 +147,15 @@ config :faktory_worker_ex,
     queues: ["github_activity"],
   ],
   start_workers: {:system, "FAKTORY_WORKERS_ENABLED", false}
+
+config :sanbase, SanbaseWeb.Graphql.ContextPlug,
+  basic_auth_username: {:system, "GRAPHQL_BASIC_AUTH_USERNAME"},
+  basic_auth_password: {:system, "GRAPHQL_BASIC_AUTH_PASSWORD"}
+
+config :ex_aws,
+  access_key_id: [{:system, "AWS_ACCESS_KEY_ID"}, :instance_role],
+  secret_access_key: [{:system, "AWS_SECRET_ACCESS_KEY"}, :instance_role],
+  region: "eu-central-1"
 
 # Import environment specific config. This must remain at the bottom
 # of this file so it overrides the configuration defined above.