+++ lib/sanbase/db_scripts/import_ico_spreadsheet.ex
@@ -6,6 +6,7 @@ defmodule Sanbase.DbScripts.ImportIcoSpreadsheet do
 
   alias Sanbase.Model.{Project, ProjectEthAddress, ProjectBtcAddress, Ico, MarketSegment, Infrastructure}
   alias Sanbase.Model.Currency
+  alias Sanbase.Model.IcoCurrencies
 
   def import(ico_spreadsheet) when is_list(ico_spreadsheet) do
     ico_spreadsheet
@@ -18,15 +19,38 @@ defmodule Sanbase.DbScripts.ImportIcoSpreadsheet do
     ico_spreadsheet_row = ico_spreadsheet_row
     |> set_infrastructure_default()
 
-    project = fill_project(ico_spreadsheet_row)
+    insert_or_update_project(ico_spreadsheet_row)
+    |> insert_or_update_ico(ico_spreadsheet_row)
+    |> insert_or_update_ico_currencies(ico_spreadsheet_row)
+  end
+
+  defp insert_or_update_project(ico_spreadsheet_row) do
+    fill_project(ico_spreadsheet_row)
     |> Ecto.Changeset.put_assoc(:market_segment, ensure_market_segment(ico_spreadsheet_row.market_segment))
     |> Ecto.Changeset.put_assoc(:infrastructure, ensure_infrastructure(ico_spreadsheet_row.infrastructure))
     |> Repo.insert_or_update!()
+  end
 
+  defp insert_or_update_ico(project, ico_spreadsheet_row) do
     fill_ico(project, ico_spreadsheet_row)
-    |> Ecto.Changeset.put_assoc(:currencies, ensure_currencies(ico_spreadsheet_row.ico_currencies))
     |> Ecto.Changeset.put_assoc(:cap_currency, ensure_currency(ico_spreadsheet_row.cap_currency))
     |> Repo.insert_or_update!()
+  end
+
+  defp insert_or_update_ico_currencies(ico, ico_spreadsheet_row) do
+    currencies = ensure_currencies(ico_spreadsheet_row.ico_currencies)
+    |> Enum.map(&Ecto.Changeset.change(&1))
+
+    Enum.each(currencies, fn currency ->
+      currency = currency
+      |> Repo.insert_or_update!()
+
+      ensure_ico_currency(ico, currency)
+      |> Repo.insert_or_update!()
+    end)
+
+    currency_ids = Enum.map(currencies, &Ecto.Changeset.get_field(&1, :id))
+    Repo.delete_all(from c in IcoCurrencies, where: c.ico_id == ^ico.id and c.currency_id not in ^currency_ids)
   end
 
   defp fill_project(ico_spreadsheet_row) do
@@ -45,10 +69,14 @@ defmodule Sanbase.DbScripts.ImportIcoSpreadsheet do
       slack_link: ico_spreadsheet_row.slack_link,
       linkedin_link: ico_spreadsheet_row.linkedin_link,
       telegram_link: ico_spreadsheet_row.telegram_link,
-      project_transparency: ico_spreadsheet_row.project_transparency,
+      project_transparency: is_project_transparency?(ico_spreadsheet_row.project_transparency),
+      project_transparency_status: ico_spreadsheet_row.project_transparency,
       team_token_wallet: ico_spreadsheet_row.team_token_wallet
       })
   end
+
+  defp is_project_transparency?(nil), do: false
+  defp is_project_transparency?(_), do: true
 
   defp fill_ico(project, ico_spreadsheet_row) do
     ensure_ico(project)
@@ -72,7 +100,7 @@ defmodule Sanbase.DbScripts.ImportIcoSpreadsheet do
 
   defp ensure_project(project_name) do
     Repo.get_by(Project, name: project_name)
-    |> Repo.preload([:eth_addresses, :btc_addresses, :market_segment, :infrastructure, {:icos, [:currencies, :cap_currency]}])
+    |> Repo.preload([:eth_addresses, :btc_addresses, :market_segment, :infrastructure, {:icos, [{:ico_currencies, [:ico, :currency]}, :cap_currency]}])
     |> case do
       result = %Project{} -> result
       nil -> %Project{}
@@ -105,7 +133,7 @@ defmodule Sanbase.DbScripts.ImportIcoSpreadsheet do
 
   defp ensure_ico(project) do
     case project do
-      %{ico: result = %Ico{}} -> result
+      %{icos: [result|_]} -> result
       _ -> %Ico{}
     end
   end
@@ -125,6 +153,18 @@ defmodule Sanbase.DbScripts.ImportIcoSpreadsheet do
         %Currency{}
         |> Currency.changeset(%{code: currency_code})
     end
+  end
+
+  defp ensure_ico_currency(ico, currency) do
+    Repo.get_by(IcoCurrencies, ico_id: ico.id, currency_id: currency.id)
+    |> Repo.preload([:ico, :currency])
+    |> case do
+      result = %IcoCurrencies{} -> result
+      nil -> %IcoCurrencies{}
+    end
+    |> Ecto.Changeset.change()
+    |> Ecto.Changeset.put_assoc(:ico, ico)
+    |> Ecto.Changeset.put_assoc(:currency, currency)
   end
 
   defp set_infrastructure_default(ico_spreadsheet_row = %IcoSpreadsheetRow{infrastructure: nil}) do