+++ priv/repo/migrations/20171215133550_rename_ico_currencies_value_to_amount.exs
@@ -1,0 +1,7 @@
+defmodule Sanbase.Repo.Migrations.RenameIcoCurrenciesValueToAmount do
+  use Ecto.Migration
+
+  def change do
+    rename table(:ico_currencies), :value, to: :amount
+  end
+end