+++ app/CHANGELOG.md
@@ -1,7 +1,22 @@
 ## CHANGELOG
 
+- 2017-26-12 | v.0.2.6:
+
+  Added Project Detailed Page
+
+- 2017-21-12 | v.0.2.4:
+
+  Update navigation menu
+  Added auth
+  Added account settings
+  Some fixes...
+  Added skeleton layout for Project Detail Page
+
 - 2017-12-12 | v.0.2.1:
 
+  Fix default sort Cashflow table
+  Fix search by project name or symbol in Cashflow page
+  Fix messageHash arg name
   Update icons.
   Added Error boundary.
   Added Raven Logs.